Finnish breaking up UISchema into files
