+++ src/applications/caregivers/config/chapters/primary/primaryHealthCoverage.js
@@ -1,19 +1,17 @@
 import fullSchema from 'vets-json-schema/dist/10-10CG-schema.json';
 import { PrimaryHealthCoverage } from 'applications/caregivers/components/AdditionalInfo';
 import { primaryCaregiverFields } from 'applications/caregivers/definitions/constants';
-import definitions from 'applications/caregivers/definitions/UIDefinitions/caregiverUI';
+import { hasHealthInsurance } from 'applications/caregivers/definitions/UIDefinitions/caregiverUI';
 
 const { primaryCaregiver } = fullSchema.properties;
 const primaryCaregiverProps = primaryCaregiver.properties;
-const { primaryCaregiverUI } = definitions;
 
 const primaryMedicalPage = {
   uiSchema: {
     'ui:description': PrimaryHealthCoverage({
       pageTitle: 'Health care coverage',
     }),
-    [primaryCaregiverFields.hasHealthInsurance]:
-      primaryCaregiverUI.hasHealthInsurance,
+    [primaryCaregiverFields.hasHealthInsurance]: hasHealthInsurance,
   },
   schema: {
     type: 'object',
