+++ src/applications/caregivers/config/chapters/secondaryOne/hasSecondaryCaregiver.js
@@ -1,8 +1,6 @@
-import definitions from 'applications/caregivers/definitions/UIDefinitions/caregiverUI';
+import { secondaryRequiredAlert } from 'applications/caregivers/definitions/UIDefinitions/caregiverUI';
 import { primaryCaregiverFields } from 'applications/caregivers/definitions/constants';
 import { SecondaryCaregiverInfo } from 'applications/caregivers/components/AdditionalInfo';
-
-const { secondaryRequiredAlert } = definitions.sharedItems;
 
 const hasSecondaryCaregiverPage = {
   uiSchema: {
